StubEngine for testing.
